% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_catchment_data}
\alias{get_catchment_data}
\title{get catchment data}
\usage{
get_catchment_data(
  catchment,
  catchment_edge_list,
  catchment_prefix = "catchment_"
)
}
\arguments{
\item{catchment}{sf data.frame NHDPlus Catchments or hyRefactor output.}

\item{catchment_edge_list}{data.frame edge list of connections
to/from catchments}

\item{catchment_prefix}{character prefix for catchment IDs}
}
\description{
get catchment data
}
