% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_catchment_edges}
\alias{get_catchment_edges}
\title{get catchment edges}
\usage{
get_catchment_edges(
  fline,
  nexus_prefix = "nexus_",
  catchment_prefix = "catchment_"
)
}
\arguments{
\item{fline}{sf data.frame NHDPlus Flowlines or hyRefactor output.}

\item{nexus_prefix}{character prefix for nexus IDs}

\item{catchment_prefix}{character prefix for catchment IDs}
}
\description{
get catchment edges
}
