% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_flowpath_data}
\alias{get_flowpath_data}
\title{get_flowpath_data
Adds flowpath realization geometry and attributes to catchments}
\usage{
get_flowpath_data(fline, catchment_edge_list, flowpath_prefix = "flowpath_")
}
\arguments{
\item{fline}{sf data.frame NHDPlus Flowlines or hyRefactor output.}

\item{catchment_edge_list}{data.frame edge list of connections
to/from waterbodies}

\item{flowpathy_prefix}{character prefix for flowpath IDs}
}
\description{
get_flowpath_data
Adds flowpath realization geometry and attributes to catchments
}
