% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hyl.R
\name{get_hydrologic_locaton}
\alias{get_hydrologic_locaton}
\title{get hydrologic location}
\usage{
get_hydrologic_locaton(hyl, flowpath)
}
\arguments{
\item{hyl}{a set of hydrologic location geometry with an identifier in the first column,
the mainstem id in the seccond column, and an sf_column in the third.}

\item{flowpath}{the flowpath table of an hygeo object.}
}
\value{
sf data.frame
}
\description{
given a set of points with a known mainstem, calculates which incremental
flowpath feature each point is on including a measure along the feature.
100 measure is the top, 0 measure is the bottom.
}
