% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_nexus}
\alias{get_nexus}
\title{get nexuses}
\usage{
get_nexus(fline, nexus_prefix = "nexus_")
}
\arguments{
\item{fline}{sf data.frame NHDPlus Flowlines or hyRefactor output.}

\item{nexus_prefix}{character prefix for nexus IDs}
}
\description{
get nexuses
}
