% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_nexus_data}
\alias{get_nexus_data}
\title{get nexus data}
\usage{
get_nexus_data(nexus, catchment_edge_list)
}
\arguments{
\item{nexus}{data.frame as returned by get_nexus}

\item{catchment_edge_list}{data.frame edge list of connections
to/from catchments}
}
\description{
get nexus data
}
