% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_waterbody_data}
\alias{get_waterbody_data}
\title{get_waterbody_data}
\usage{
get_waterbody_data(fline, waterbody_prefix = "waterbody")
}
\arguments{
\item{fline}{sf data.frame NHDPlus Flowlines}

\item{waterbody_prefix}{character prefix for waterbody IDs}
}
\description{
get_waterbody_data
}
