% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_waterbody_edge_list}
\alias{get_waterbody_edge_list}
\title{get waterbody edge list}
\usage{
get_waterbody_edge_list(catchment_edge_list,
  catchment_prefix = "catchment", waterbody_prefix = "waterbody")
}
\arguments{
\item{catchment_edge_list}{data.frame as returned by get_catchment_edges}

\item{catchment_prefix}{character prefix for catchment IDs}

\item{waterbody_prefix}{character prefix for waterbody IDs}
}
\description{
get waterbody edge list
}
