% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_catchment_data}
\alias{get_catchment_data}
\title{get catchment data}
\usage{
get_catchment_data(catchment, catchment_prefix = "catchment")
}
\arguments{
\item{catchment}{sf data.frame NHDPlus Catchments}

\item{catchment_prefix}{character prefix for catchment IDs}
}
\description{
get catchment data
}
