% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_nhd_crosswalk}
\alias{get_nhd_crosswalk}
\title{Get NHD Crosswalk}
\usage{
get_nhd_crosswalk(x, catchment_prefix = "catchment_")
}
\arguments{
\item{x}{sf data.frame output from reconcile_collapsed_flowlines() function}

\item{catchment_prefix}{character prefix to be appended to local_id output.}
}
\description{
Get NHD Crosswalk
}
