% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_waterbody_edge_list}
\alias{get_waterbody_edge_list}
\title{get waterbody edge list}
\usage{
get_waterbody_edge_list(fline, waterbody_prefix = "wat-")
}
\arguments{
\item{fline}{sf data.frame NHDPlus Flowlines or hyRefactor output.}

\item{waterbody_prefix}{character prefix for waterbody IDs}
}
\description{
Takes flowlines from either NHDPlus or hyRefactor and returns
dendritic waterbody topology. This maps onto the HY_Features waterbody class
and downstream waterbody associations.
See: https://docs.opengeospatial.org/is/14-111r6/14-111r6.html#_the_hydrographic_network_model
}
