% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_hygeo.R
\name{write_hygeo}
\alias{write_hygeo}
\title{Write hygeo}
\usage{
write_hygeo(
  hygeo_list,
  out_path,
  edge_list_format = "json",
  data_format = "geojson",
  edge_map = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{hygeo_list}{list of class hygeo containing:
catchment, flowpath, nexus, catchment_edges, and waterbody_edges.}

\item{out_path}{character path to store outputs}

\item{edge_list_format}{character 'json' or 'csv'}

\item{data_format}{character 'geojson' or 'gpkg'}

\item{edge_map}{logical write edge map json file"}

\item{overwrite}{logical overwrite output or not}
}
\description{
Write hygeo
}
